+++ CHANGELOG.md
@@ -10,18 +10,33 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - New colors (#474)
 - New typography (#475)
+- Ability to create `HTML` and `CSS` files in a `HTML` project (#478)
+- Help text in the new file modal (#478)
+
+### Changed
+
+- Handling of internal/external links for HTML/CSS projects (#483)
+- Position run button in separate bar (#471)
+
+### Fixed
+
+- Updated Sentry to fix compile-time error
+- Ability to run HTML projects by clicking run button (#481)
 - Add dev-container support (#489)
 - Distinguish icon type (#470)
 
 ### Changed
 
 - Updated spacing variables to match the new design system (#460)
+- Position run button in separate bar (#471)
 - Color usage (#474)
+
 
 ### Fixed
 
 - Modal button alignment support (#460)
 - Storybook setup and dev deployment (#461)
+- HTML auto-run functionality (#481)
 - Updating Sentry to fix compile-time error (#487)
 - Fix local package build process (#488)
 - Color naming convention (#474)